Refactor Supabase client imports and update resume upload
Renamed 'auth-client.ts' to 'client.ts' in the server core database package and updated all imports accordingly in both CXC and Web apps. Refactored the resume upload API to use the new client creation functions and improved error handling and validation.
