+++ apps/cxc/app/api/applications/resumes/route.ts
@@ -1,17 +1,21 @@
 import { NextRequest } from "next/server";
-import { AuthService } from "@uwdsc/server/core/services/authService";
-import { supabase } from "@uwdsc/server/core/database/client";
+import { createAuthService } from "@/lib/services";
+import { createSupabaseBrowserClient } from "@uwdsc/server/core/database/client";
 
 const MAX_BYTES = 10 * 1024 * 1024; // 10 MB
 const ALLOWED_MIME = new Set([
   "application/pdf",
   "application/vnd.openxmlformats-officedocument.wordprocessingml.document", // .docx
-  "application/msword", 
+  "application/msword",
 ]);
 
 function extFromMime(mime: string) {
   if (mime === "application/pdf") return "pdf";
-  if (mime === "application/vnd.openxmlformats-officedocument.wordprocessingml.document") return "docx";
+  if (
+    mime ===
+    "application/vnd.openxmlformats-officedocument.wordprocessingml.document"
+  )
+    return "docx";
   if (mime === "application/msword") return "doc"; // optional
   return null;
 }
@@ -20,14 +24,21 @@ export async function POST(req: NextRequest) {
   try {
     const form = await req.formData();
     const file = form.get("file") as File | null;
+    const supabase = createSupabaseBrowserClient();
 
     if (!file) {
-      return Response.json({ error: "Missing 'file' upload field" }, { status: 400 });
+      return Response.json(
+        { error: "Missing 'file' upload field" },
+        { status: 400 }
+      );
     }
 
     // Validate size
     if (file.size > MAX_BYTES) {
-      return Response.json({ error: `File too large. Max ${MAX_BYTES / (1024 * 1024)} MB.` }, { status: 413 });
+      return Response.json(
+        { error: `File too large. Max ${MAX_BYTES / (1024 * 1024)} MB.` },
+        { status: 413 }
+      );
     }
 
     // Validate mime type
@@ -39,7 +50,7 @@ export async function POST(req: NextRequest) {
       );
     }
 
-    const authService = new AuthService();
+    const authService = await createAuthService();
     const { user, error: userErr } = await authService.getCurrentUser();
 
     if (userErr || !user) {
@@ -48,7 +59,10 @@ export async function POST(req: NextRequest) {
 
     const ext = extFromMime(mime);
     if (!ext) {
-      return Response.json({ error: "Unsupported content type" }, { status: 415 });
+      return Response.json(
+        { error: "Unsupported content type" },
+        { status: 415 }
+      );
     }
     const objectKey = `${user.id}/${file.name}`;
 
@@ -61,16 +75,17 @@ export async function POST(req: NextRequest) {
       });
 
     if (uploadErr) {
-      return Response.json({ error: `Upload failed: ${uploadErr.message}` }, { status: 500 });
+      return Response.json(
+        { error: `Upload failed: ${uploadErr.message}` },
+        { status: 500 }
+      );
     }
-
 
     return Response.json({
       message: "Upload successful",
-      key: objectKey
+      key: objectKey,
     });
-  } 
-  catch (err: any) {
+  } catch (err: any) {
     console.error("Resume upload error:", err);
     return Response.json({ error: "Internal server error" }, { status: 500 });
   }
